// src/app/modules/campaign-setting/setting-point-incentive/models/GoalParameterListModel.ts
export interface GoalParameterListModel{
    goalParameterRangeConfigurationId: number
    campaignSettingId?: number
    currencyId?: number
    currencyName: string
    rangeNo?: number
    pointAmount?: number
    rangeFrom?: number
    rangeTo?: number
    createdBy?: string
    createdDatetime?: string
    updatedBy?: string
    updatedDate?: string
}